[BE] 질문 기능 수정, 추가 (#43)
* ✨feat: 질문 관련 엔티티, dto 추가

* ✨feat: 질문 등록, 수정, 삭제, 조회 기능 추가

* ✨feat: 공통 기능 변경

* ✨feat: 질문 Dto 유효성 검증 추가

* etc: Tag entity 삭제 - questionTag 통합

* ✨feat: 질문 조회수 기능 추가, 예외처리

* ✨feat: 질문 조회 시 답변 함께 출력하는 기능 추가

* ✨feat: 답변 관련 엔티티, dto 추가

* ✨feat: 답변 등록, 수정, 삭제, 조회 기능 추가

* 🐛fix: 질문 수정 시 태그 중복 등록 수정

* ✨feat: ExceptionCode 추가

* ✨feat: 회원 비밀번호 수정 기능 , 로그인한 회원 조회 기능 추가

* ✨feat: 응답 데이터 수정

* ✨feat: 질문, 답변, 태그 추가 매핑

* ✨feat: 질문, 답변 추천-비추천 기능 추가

+++ server/stackoverflow/src/main/java/com/y2k/stackoverflow/answer/service/AnswerService.java
@@ -1,9 +1,14 @@
 package com.y2k.stackoverflow.answer.service;
 
 import com.y2k.stackoverflow.answer.entity.Answer;
+import com.y2k.stackoverflow.answer.entity.AnswerVote;
 import com.y2k.stackoverflow.answer.repository.AnswerRepository;
+import com.y2k.stackoverflow.answer.repository.AnswerVoteRepository;
 import com.y2k.stackoverflow.exception.BusinessLogicException;
 import com.y2k.stackoverflow.exception.ExceptionCode;
+import com.y2k.stackoverflow.member.entity.Member;
+import com.y2k.stackoverflow.member.service.MemberService;
+import com.y2k.stackoverflow.question.entity.Question;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,9 +18,13 @@ import java.util.Optional;
 public class AnswerService {
 
     private final AnswerRepository answerRepository;
+    private final AnswerVoteRepository answerVoteRepository;
+    private final MemberService memberService;
 
-    public AnswerService(AnswerRepository answerRepository) {
+    public AnswerService(AnswerRepository answerRepository, AnswerVoteRepository answerVoteRepository, MemberService memberService) {
         this.answerRepository = answerRepository;
+        this.answerVoteRepository = answerVoteRepository;
+        this.memberService = memberService;
     }
 
     public Answer createAnswer(Answer answer) {
@@ -25,6 +34,11 @@ public class AnswerService {
     public Answer updateAnswer(Answer answer) {
         // 해당 답변이 존재하는 지 검증 한 후
         Answer findAnswer = findVerifiedAnswer(answer.getAnswerId());
+        //로그인 유저와 답변 작성 유저와 비교해서 같다면 수정, 아니라면 접근 금지 예외 발생
+        if(findAnswer.getMember().getMemberId() != memberService.getLoginMember().getMemberId()) {
+            throw new BusinessLogicException(ExceptionCode.ACCESS_FORBIDDEN);
+        }
+
         // 답변 내용 업데이트
         Optional.ofNullable(answer.getContent())
                 .ifPresent(content -> findAnswer.setContent(content));
@@ -36,8 +50,16 @@ public class AnswerService {
         return answerRepository.findAll();
     }
 
+    public List<Answer> findAnswersQuestion(Question question) {
+        return answerRepository.findAllByQuestion(question);
+    }
+
     public void deleteAnswer(Long answerId) {
         Answer findAnswer = findVerifiedAnswer(answerId);
+        //로그인 유저와 답변 작성 유저와 비교해서 같다면 삭제, 아니라면 접근 금지 예외 발생
+        if(findAnswer.getMember().getMemberId() != memberService.getLoginMember().getMemberId()) {
+            throw new BusinessLogicException(ExceptionCode.ACCESS_FORBIDDEN);
+        }
         answerRepository.delete(findAnswer);
     }
 
@@ -46,4 +68,56 @@ public class AnswerService {
         Answer findAnswer = optionalAnswer.orElseThrow(() -> new BusinessLogicException(ExceptionCode.ANSWER_NOT_FOUND));
         return findAnswer;
     }
+
+    /**
+     * 한 질문에 이미 추천 or 비추천한 사람이면 다시 추천 못함
+     * answer_vote 테이블에서 로그인 유저 memberId로 조회 후, 결과가 존재하지 않으면 추천 가능
+     */
+
+    public Answer likeAnswerVote(Long answerId, Member member) {
+        Answer findAnswer = findVerifiedAnswer(answerId);
+
+        // 로그인 한 회원이 추천 눌렀는지 확인 후,
+        // 안눌렀다면 answer_vote 테이블에 answer_id와 member_id를 넣어 중복 방지
+        if(findVerifiedVoteMember(answerId)) {
+            AnswerVote answerVote = new AnswerVote();
+            answerVote.setAnswer(findAnswer);
+            answerVote.setMember(member);
+            answerVoteRepository.save(answerVote);
+
+            //추천 누른 질문에 전체 추천 수 카운트
+            findAnswer.setVotes(findAnswer.getVotes() + 1);
+        }else {
+            throw new BusinessLogicException(ExceptionCode.ACCESS_FORBIDDEN);
+        }
+        return answerRepository.save(findAnswer);
+    }
+
+
+    public Answer dislikeAnswerVote(Long answerId, Member member) {
+        Answer findAnswer = findVerifiedAnswer(answerId);
+
+        // 로그인 한 회원이 추천 눌렀는지 확인 후,
+        // 안눌렀다면 answer_vote 테이블에 answer_id와 member_id를 넣어 중복 방지
+        if(findVerifiedVoteMember(answerId)) {
+            AnswerVote answerVote = new AnswerVote();
+            answerVote.setAnswer(findAnswer);
+            answerVote.setMember(member);
+            answerVoteRepository.save(answerVote);
+
+            //추천 누른 질문에 전체 추천 수 카운트
+            findAnswer.setVotes(findAnswer.getVotes() - 1);
+        }else {
+            throw new BusinessLogicException(ExceptionCode.ACCESS_FORBIDDEN);
+        }
+        return answerRepository.save(findAnswer);
+    }
+
+    //로그인 한 회원이 추천 눌렀는지 확인해서 안 눌렀으면 TRUE 리턴
+    private boolean findVerifiedVoteMember(Long answerId) {
+        Answer findAnswer = findVerifiedAnswer(answerId);
+        return answerVoteRepository.findByAnswerAndMember(findAnswer, memberService.getLoginMember()).isEmpty();
+    }
+
+
 }
